make other modules, need to implement two more routines for posix support
